Move image sources to separate module
