+++ src/components/actions/Actions.jsx
@@ -1,13 +1,15 @@
 import { Fragment } from "react";
 import Button from "src/components/button/Button";
 import {
-  ICON_MARK,
-  ICON_PEN,
-  ICON_BIN,
   ALT_MARK_TEXT,
   ALT_EDIT_TEXT,
   ALT_DELETE_TEXT,
 } from "src/utils/constants/constants";
+import {
+  ICON_MARK,
+  ICON_PEN,
+  ICON_BIN,
+} from "src/utils/constants/imageSources";
 
 const Actions = () => {
   return (
