+++ src/store/actions/taskActions.js
@@ -1,5 +1,5 @@
 import { ADD_TASK } from "src/store";
-import createDate from "src/utils/helpers/createDate";
+import { createDate } from "src/utils/helpers/utilities";
 
 export const addTask = (title) => {
   return {
