+++ src/utils/constants/constants.js
@@ -1,19 +1,3 @@
-import markIcon from "src/assets/images/mark.png";
-import penIcon from "src/assets/images/pen.png";
-import binIcon from "src/assets/images/bin.png";
-import logo from "src/assets/images/logo.png";
-import navLogo from "src/assets/images/nav-logo.png";
-import searchIcon from "src/assets/images/search.png";
-import addIcon from "src/assets/images/add.png";
-
-export const ICON_MARK = markIcon;
-export const ICON_PEN = penIcon;
-export const ICON_BIN = binIcon;
-export const LOGO = logo;
-export const NAV_LOGO = navLogo;
-export const ICON_SEARCH = searchIcon;
-export const ICON_ADD = addIcon;
-
 export const ALT_MARK_TEXT = "Mark task as done";
 export const ALT_EDIT_TEXT = "Edit task";
 export const ALT_DELETE_TEXT = "Delete task";
