--- src/utils/helpers/sanitizeInput.js
@@ -1,5 +1,0 @@
-function sanitizeInput(input) {
-  return input.replaceAll(/<\/?[^>]+(>|$)/gi, "").trim();
-}
-
-export default sanitizeInput;
