+++ src/utils/helpers/utilities.js
@@ -1,8 +1,10 @@
-const createDate = () => {
+export const createDate = () => {
   const currentDate = new Date();
   return `${currentDate.getDate()}.${currentDate.getMonth() + 1}.${String(
     currentDate.getFullYear()
   ).slice(-2)}`;
 };
 
-export default createDate;
+export const sanitizeInput = (input) => {
+  return input.replaceAll(/<\/?[^>]+(>|$)/gi, "").trim();
+};
